+++ mayan/apps/platform/dependencies.py
@@ -7,10 +7,10 @@ PythonDependency(
     module=__name__, name='greenlet', version_string='==3.0.3'
 )
 PythonDependency(
-    module=__name__, name='gunicorn', version_string='==21.2.0'
+    module=__name__, name='gunicorn', version_string='==22.0.0'
 )
 PythonDependency(
-    module=__name__, name='sentry-sdk', version_string='==1.41.0'
+    module=__name__, name='sentry-sdk', version_string='==1.45.0'
 )
 PythonDependency(
     module=__name__, name='whitenoise', version_string='==6.6.0'