+++ mayan/apps/storage/dependencies.py
@@ -4,13 +4,13 @@ PythonDependency(
     module=__name__, name='boto3', version_string='==1.33.7'
 )
 PythonDependency(
-    module=__name__, name='django-storages', version_string='==1.14.2'
+    module=__name__, name='django-storages', version_string='==1.14.3'
 )
 PythonDependency(
     module=__name__, name='extract-msg', version_string='==0.48.5'
 )
 PythonDependency(
-    module=__name__, name='google-cloud-storage', version_string='==2.15.0'
+    module=__name__, name='google-cloud-storage', version_string='==2.16.0'
 )
 PythonDependency(
     module=__name__, name='pycryptodome', version_string='==3.20.0'